VLM wing: CPACS: compensate variable relative airfoil thickness
The relative airfoil thickness changes the relative camber
which changes the zero lift angle of attack.
The local wing incidence angle is now adjusted depending
on the relative airfoil thickness. The reference relative
airfoil thickness is at 50% dimensionless span coordinate.
Moreover, the definition of the wing twist was adjusted.
The local wing twist is now always defined with respect
to the wing root.